feat(geometry): the analytic B-rep is double precision
The float coordinate ceiling, decided in principle in design-decisions.md and now executed
for the B-rep. Positions, curves, surfaces, edge parameters and radii are double; the mesh
and the renderer stay single precision, and toMesh is the boundary where the conversion
happens.

WHAT MADE IT ONE CHANGE RATHER THAN 400. The B-rep namespace named its vector through a
single alias, `using nexus::render::Vec3;`. Pointing that alias at a new `Vec3d` migrated
every stored point and every internal computation in AnalyticBRep, BRepBoolean and
BRepSurfaceIntersect at once. Vec3d converts from render::Vec3 IMPLICITLY, because widening
cannot lose anything -- which is why 46 of the 48 test files that use brep types needed no
change beyond dropping a `using nexus::render::Vec3;` that shadowed the alias. Narrowing is
`toFloat()` and explicit, so every place precision is given up is visible.

THE PREDICATES GOT BETTER, not just compatible. RobustPredicates now takes the double vector
types; the float ones widen into them, so every existing call is unchanged and exact. The
implementation already converted to double on entry -- what changes is that it is no longer
handed a value that was rounded to float first. The expansion arithmetic was always exact for
arbitrary doubles; now it is given some.

SERIALIZATION v4 writes positions, radii and edge parameters as double, and reads v1-v3 as
float. The old test of that rule relabelled a CURRENT blob with an older version byte, which
only worked while every version shared one payload layout -- across a width change it would
be a v4 payload with a v1 header. Replaced with a genuine v3 blob produced by the v3 writer
before the migration and committed as tests/kernel/fixtures/brep_v3_box.nxb, which cannot
drift with the code. It decodes to the same solid the current writer produces, so the
widening moved no geometry; it only stopped rounding it.

ONE REAL REGRESSION, found and fixed. Body::transform narrowed the point through a float
Mat4 -- so translating a cylinder by (0,0,0) perturbed its coordinates enough that it no
longer compared coincident with an untranslated copy, and the identity boolean of two
identical cylinders returned empty. The matrix's ENTRIES are single precision but the point
must not be: promoted and multiplied in double.

Two tests were fixture artifacts rather than failures: SegmentTriangleBeatsNaiveFloat's
"naive float32" comparison had silently become naive DOUBLE (and therefore accurate, turning
the assertion into a tautology) -- it is now explicitly render::Vec3; and the non-finite
serialization probe poked a 4-byte Inf pattern at an offset that now holds a double.

MEASURED. A coordinate needing more than float's 24-bit mantissa now survives a round trip
exactly: 1.2345678901234567 reads back unchanged where float gives 1.2345678806304932. That
is the ceiling lifting, and it is asserted.

WHAT THIS DOES NOT YET BUY, stated because the number is easy to misread. End-to-end
geometric accuracy is unchanged at ~6e-8 relative, because the PRIMITIVE BUILDERS still
place their vertices with float trig -- a double vertex holding a float-accurate point is
still float-accurate. Widening makeCylinder alone was tried and REVERTED: the cylinder
improved (5.9e-8 -> 4.4e-8) and the box got WORSE (1.24e-7), because a double-built
cylinder's seam vertices agree with a float-built box's less well than two float-built ones
did. The primitives migrate together or not at all, and that is the next stage.

Test: ctest --test-dir build -> 2495/2495 (5 Vulkan-HW skipped). API-freeze audit clean --
no header added, so the manifest is untouched.

+++ apps/Nexus-Modeling/src/kernel/src/geometry/BRepSurfaceIntersect.cpp
@@ -96,7 +96,7 @@ SurfaceIntersection planeSphere(const Surface& plane, const Surface& sphere, Tol
         r.point = foot;
         return r;
     }
-    const float cr = std::sqrt(std::max(0.f, sphere.radius * sphere.radius - d * d));
+    const double cr = std::sqrt(std::max(0.0, sphere.radius * sphere.radius - d * d));
     r.kind = SurfaceIntersectionKind::Circle;
     r.curve = circleCurve(foot, n, cr);
     return r;
